Simplified agent, server and web conf templates
Eliminated the verbosely commented original configuration files in favor
of a more compact form. This way we don't need to keep track of most of
the changes that could occur in new versions of Zabbix.

For the same reason we keep just some basic parameters in those conf
files as pillars, allowing compatibility with 2.0.x, 2.2.x and 2.4.x. If
some other parameters are need, it's possible to:

a) use the extra_conf pillar value
b) override the template