Deleted args overriding saltstack mysql conf ...
and eliminated requirement of a mysql service in the minion.

Even if in my original formula there was a `service: mysql-server`
requisite for the `zabbix_db: mysql_database_present` state, I'm willing
to eliminate that and add a `host`parameter. Zabbix requires a MySQL
service that could be in a different host (not even a minion), so no
need to set that dependence. If a MySQL server is required then it
should be done with the mysql formula AND orchestrated conveniently.

The `connection_user` and `connection_pass` add repetitive lines and sets a
configuration that is designed to be in /etc/salt/minion
(http://docs.saltstack.com/en/latest/ref/modules/all/salt.modules.mysql.html)
in order to facilitate the general mysql management from SaltStack.
Actually, in some distros, root is allowed to manage MySQL, so no need
to configure anything in those cases. As stated at
http://docs.saltstack.com/en/latest/ref/states/all/salt.states.mysql_user.html,
the MySQL authentication information specified in the minion config file
_can be overridden_ by some arguments, but I don't think that's the
general case that formulas should represent.