Specifically disable Recommends for zabbix-server
The code removed in this commit was written to ensure that the
installation of the zabbix-server package does not pull in the
mysql-server package via Recommends on Debian. The way it was written
did, however, disable the installation of *any* recommended package
globally and outwith the scope of this formula.

One problematic aspect of this commit is that it requires users to use
at least salt 2015.5.0 as setting "install_recommends" was not supported
before.

This fixes #19