Be type agnostic when referencing version_repo
The comparison to zabbix.version_repo _happened_ to work if the value
in the pillar was a string. If the value in the pillar was not a string,
the comparison would fail. So here, we take version_repo, force it to a
float, and compare it to a float. In this way, it doesn't matter if the
repo value is '3.2' or 3.2; the comparison works correctly either way.